Update README.md
Motivated and enthusiastic Junior Front end Developer with a strong passion for web development and a solid foundation in HTML, CSS, and JavaScript. Adept at collaborating with cross-functional teams to create engaging and responsive user interfaces. Dedicated to continuous learning and staying up-to- date with the latest industry trends and technologies. Seeking an entry-level position to contribute to innovative projects and further enhance technical skills.